+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased -->
 
+## 1.0.3 - 2021-08-11
+
+- [Added link to daac name selected in questions](7a7e7b3e80075266272b8e323fd3eb45fed55827)
+- [Added daac name selected to questions (being set from mixin in setting the active location)](c752ac0373b0561d0bf4381bb83d0a4f2b8ca075)
+
+## 1.0.2 - 2021-08-10
+
+- [Updated jsons for publication_dois to be table format for multirow](26bfbd65dc6c514aebbc833493eb254e1442affa)
+
 ## 1.0.1 - 2021-07-22
 
 - [Updated jsons from api and moved them to public folder](e5f2359446313fde12120b6bb19918f01f31c8d8)